Allow plugins to add and manipulate Yum's configuration file options in the same way as they are defined in Yum. This means that plugin option defines are far more flexible and powerful than before. The old API (registerOpt) still exists but is deprecated (with a saner implementation). I will update the wiki with info on how plugins can now define their own options.
This cleans up a lot of plugin related code but changes the Yum API slightly
wrt to initialisation. Documentation about this to come on yum-devel.

This commit also includes Misa's (misa+yum@redhat.com) patch to add a config
option for specifying the directories to look in for plugin configuration
files.