fix up YumAvailablePackageSqlite.returnSimple() - if you select %s and pass it in on the cursor.execute with sqlite you end up with a quoted string. So "select %s from foo", ('somevar')
results in this:

select 'somevar' from foo

which means EVERY answer you get back will be 'somevar' and not the value in
the row for that column.

If anyone knows how to turn off that quoting per-variable in sqlite, let me
know - the docs didn't have it.

We also needed an simple-dict to db column-name map so querying for certain
things would return the right information.