Remove _checkUpdate
If a transaction member was marked as 'update', we'd run _checkInstall on it,
then run _checkRemove on all of the packages that it was updating. After this,
we'd then run _checkRemove on all transaction members marked as 'updated'. So
we were running _checkRemove on every 'updated' package twice. Get rid of this.