From Gijs Hollestelle --
I managed to trace it to the already_seen cache in depsolver.py it
turns out that  when resolveDeps has missing requirements and restarts
the depsolving loop the packages that caused that missing requirements
are marked as already seen in the cache and their dependencies not
always resolved again, this causes problems at least in the case of an
upgrade but I can imagine that it could cause other problems as well.

I've included a small patch to fix this, it removes all packages
resulting in missing dependencies from the already seen cache before
starting the next round of depsolving.