commit patch from Florian Festi:
"The attached patch makes sure that an updated pkg is removed even if the
updating pkg already got added to the transaction. This fixes updates from
multilib (32 and 64 bit installed) to noarch. Currently the first update
operation (e.g. 32bit -> noarch) works but the second (64bit -> noarch) is
ignored. As result there are two pkg in different arches and versions which
will most likely result in an multilib file conflict. The current patch
makes sure that the updated pkg is only rejected if itself is already in the
transaction (as a REMOVE)."