Create an atomic MD update code path:
 Pros.
 -----

 Should stop these metadata update problems:

  1. We get corrupted comps/etc. files on the master and everyone has
problems.
  2. We hit mirror(s) that have an updated repomd.xml but nothing else.
  3. We don't have a network but the cache does a timeout and urlgrabber
kills repomd.xml and we can't get a new one (makes yum stop working).

 Should stop "back in time updates". Ie. we hit an old mirror, and we
basically go back in time for updates.

 Helps stop yum cmd line usage hitting network. Basically yum-updatesd
can now download all of MD data we need and not just primary.xml so we
don't have a problem where user does "yum blah" which happens to need an
MD file we don't have so we hit the network.
 This actually has follow-on problems where the file isn't the same anymore
but we haven't updated repomd.xml yet (or the network is down, think yum
deplist /usr/bin/foo).

 Cons.
 -----

 Downloads more stuff at once. Basically the current yum model only ever
downloads what we need, when we need it, now we'll download more of the MD
files whenever repomd.xml gets updated and they need it. However we can
config. to do the old behaviour (and/or how much of the MDs we get).

 _If_ we don't have a full set of MD currently, and we fail to get a new
batch of data then we'll revert back to the non-full set. So it's
_possible_ that we'll need one of the files in the set we don't
have, but that would be available if we had used the traditional code
path (Ie. the error was in one of the files we don't currently need).
 This is a very weird edge case, and if you care a good solution IMO
is to always use group:all ... then we always have a complete set.

 Minor CPU speed hit, due to parsing two sets of repomd.XML.