Add some obsoletes tests for the git metapackage rename
I was all ready to blame yum for my inability to update git in F8 to the
new and renamed version in rawhide, so much so that the first thing I
did was write these tests, which would surely fail and lead me to some
obscure bug in yum. I would fix the bug and everyone would be happy.

Well the version obsoletes I put in git's spec was off by one. So here
are the tests, which will serve as a testament to yum's awesomeness, and
my carelessness.