correct searchFiles for primary file search
when we performing searchFiles() against the sqlitesacks we were not checking
if the file names were from the set of files that the primarydb has already.
So we were downloading the filelistsdb in a bunch of cases where we simply
didn't need to. This fixes the simple case of: there are no globs and the
file is a primary. It also marks a fixme to be finished that there are
globs but the globs matches up to a primarydb file. I'll get to that one shortly.