Add transactionCacheFileRequires() and fileRequiresData()
 This is the most scarey part of the rpmdb caching. I think it's good.
But it's somewhat complicated, and if we screw it up (it isn't in sync.
with the rpmdb) bad things will likely result.
 We are protected from rpm doing crazy things due to the way
transactionResultVersion() works, so all that matters is that our
logic is right for what is in the transaction.
 Both installedFileRequires and installedFileProviders should always
be updated for the transaction, and thus. be identical to rm'ing the
cache.

 And I've tested it, and it seems to work :).