pkgtags db - searching
this patch adds the pkgtags metatadata and the integrates it into our
search mechanism. Completely optional repodata format.

Also adds a decompress() function from yum.misc that allows for us
to pass files to it and not care what compression mechanism they are using
it's a bit simplistic at the moment but can easily be enhanced.