fix for https://bugzilla.redhat.com/show_bug.cgi?id=613399 in two ways - so that various api callers get
the benefit no matter which objects they are using.

If we have a mediaid and no mediafunc and no other urls to use - then mark the repo as 'skip_if_unavailable' and 'disabled'
Then if we get any repo obj back from repo.setup() as not enabled then mark it as disabled in the repos object - this is
mostly so our plugins don't accidentally trample all over it.