add support for saving out depsolved transactions and reloading them.
This saves a considerable amount of time if you find you have to deal with
gpg keys, between updates. It is also flexible enough that it could allow users
with systems in relative, or precise, lockstep, to pass around a .ts file and
have the deps pre-solved - cutting out a considerable amount of time searching
for providers.

this adds the load-ts/load-transaction command to yum-cli which lets you load them from a file