Revert "Don't keep full headers in memory during transaction"
This reverts commit 3302369486dc07edc2428c616ed2b8843562f615.
The callback key isn't as contained within the callback classes
as I'd hoped, various places are testing isinstance(h, rpm.hdr) etc.
Since changing the callback style is an incompatible change, it needs
to be an opt-in feature enabled by callers who are ready for it. And
since its going to be incompatible, we can do better than copy data
around.