Add a new opt-in callback mode where only txmbr is used as the key
We really only want the txmbr in the key, everything else can be
reached through it. This avoids copying any data around as txmbrs
reside in memory anyway, so the key is just a cheap reference
to what we already have. With the new key -> txmbr helper in the
callback, this is truly trivial to handle there. The new mode
makes life in the callback somewhat easier and saves gobs of memory.