Workaround rpm chroot insanity, so we can use sqlite from unInstStop() cb.
 So the problem is that AGAIN, when we log stuff for installed rpms we
do it in _instCloseFile, which is outside the chroot ... but when we log
stuff for erase rpms we do it in _unInstStop, which is inside the
chroot.
 sqlite will open it's journal file, which is it has a working path for
from outside the chroot ... being inside the chroot breaks this, which
results in complete failure of everything.

 Fix is to move out of the chroot, before we call the history/sqlite APIs, and
then move back in again ... basically what rpm does.