Fix possible endless loop in yum-cron
In the unlikely but possible case where LOCKDIR contained files _other_ than
the PIDFILE being tested for, rmdir wouldn't remove it, causing the lock to
persist and then rapidly be tested over and over as the script respawns.

Now uses rm -rf to zap the LOCKDIR (as is done at other points in the code).