Sick hack to work around python readline changes. BZ 448864
 This is a modification of a patch James posted a while ago, see
http://lists.baseurl.org/pipermail/yum-devel/2010-September/007495.html
 The changes from that patch are: changing __builtin__.raw_input so
that _sick_hack_raw_input is used when raw_input is called from other
modules; calling to_utf8 on the prompt to raw_input, so that non-ascii
prompts don't break anything; and finally resetting sys.stdout fewer
times, to be simpler.  I've tested the shell quite a bit using a
Japanese locale　so that lots of unicode characters are printed, I
don't speak Japanese, so it might be outputting nonsense, but as far
as I can tell everything works as it should.