packaging: moving stuff around.
we can not have another /usr/lib/pyton2.7/site-packages/yum, that would
clash with the yum package. in the new arrangement everything will be
placed under site-packages/dnf. this will make layout of the git repo same
as the layout of installed package where the files move as follows:

/usr/share/yum-cli -> /usr/lib/python2.7/site-packages/dnf/yum-cli
/usr/lib/python2.7/site-packages/yum -> /usr/lib/python2.7/site-packages/dnf/yum
/usr/bin/yum -> /usr/bin/dnf

config files will be separated later too.