correct the _repoXML.srcfile value.
In a scenario where we consider metadata expired according to our
metadata_expire setting yet we verify it is still valid according to the
latest metalink we can naturally use the old metadata. yum created the
RepoMD object from a copy of the original metadata and never bothers to
change its .srcfile back, even after the .old.tmp file is cleaned up. dnf
however needs an existing .srcfile to pass it hawkey (yum probably was
never interested in the file name at this point as it already had a valid
RepoMD object then).

This fixes assert failure in hawkey in sack.c.