doc: add foundation documenation system files.
DNF will go with Sphinx just like Hawkey. This adds the conf.py and
skeleton templates and tweaks build system and the RPM spec so that
manpage is a part of the distributed package.

Note that conf.py imports dnf.const and so the module has to be importable
during build time.