packaging: python-hawkey is now a build require.
With the current generation Sphinx there is no way to tell [1] we don't need to
include the api documentation for manpage build. So the api *.rst sources are
always scanned and dnf modules are always imported. Therefore, for the man page
build not to fail, the hawkey package must be importable too.

[1] https://groups.google.com/forum/?fromgroups=#!topic/sphinx-dev/Zrh7T5y4FVY