tests: use iniparse.compat instead of configparser.
Using iniparse.compat consistently everywhere saves us a bit of a py2/3 compat
issues (the module renamed between the versions), plus we have no reason to
assume the standard lib implementation is better. should this change
(i.e. someone measures the std lib is faster) we can drop iniparse in a blink of
an eye.