Integrate DemandSheet into CLI.
Built-in commands can keep on using their class attributes as a
convenience for now or define configure(). Only the cli.demand approach is
an API though. It is more flexible, allowing commands to demand features
dynamically based on command line argumens. Also, DemandSheet does
checking that once demanded feature is not disabled again.