dnf-makecache.service: start service after akmods-shutdown.service (RhBug:1187111)
Another situation to consider is shutting down the system at the
moment dnf-makecache.service is running. There is a unit
akmods-shutdown.service which builds new modules at shutdown time;
it is implemented as an oneshot service with a no-op start and actual
work on stop.

So, when a system starts to shut down in such circumstances, a stop
job will be enqueued for akmods-shutdown.service and the existing
start job for dnf-makecache.service will be reversed into a stop job.

Thus, because in systemd ordering at stop time is inverse to that which
is specified, order dnf-makecache.service after akmods-shutdown.service.

Reference: https://bugzilla.redhat.com/show_bug.cgi?id=1187111
Signed-off-by: Ivan Shapovalov <intelfx100@gmail.com>
Signed-off-by: Igor Gnatenko <i.gnatenko.brain@gmail.com>