clean: operate on all cached repos (RhBug:1278225)
Previously, the clean command would only include the enabled repos with
the current $releasever in the cleanup.  In order to also include any
leftovers from previous dnf runs with --releasever or --repofrompath,
the user had to specify those options again.  That was not intuitive and
caused confusion in some scenarios.  For example, one would expect that,
after running these commands on Fedora 22, the data fetched by
system-upgrade would be gone:

$ dnf system-upgrade download --releasever=23
$ dnf clean all

However, it wouldn't.  Only running the clean command with
--releasever=23 would accomplish that.  The same applies to the
--repofrompath option.

This commit redefines the clean command so that it removes all the data
of the specified type from the cache, whether or not it belongs to a
configured repo.