build: place configured files in build tree
Don't pollute the source tree with built files from out-of-source builds
(oosb).  Otherwise it wouldn't be possible to have multiple builds with
different configurations.

To ensure that, with oosb, the binaries continue to work for developers
without installing, we need to configure the dnf/__init__.py file so
that it's placed in CMAKE_BINARY_DIR and extend its __path__ attribute
so that the source modules are found.  To ensure that the configuration
doesn't happen with such builds meant to be installed, the USE_CHECKOUT
cmake option has been introduced.