RepoPersistor: only write to disk when requested
Allow for staging changes to a RepoPersistor instance before actually
writing them to disk by introducing the save() method.  This will be
useful as soon as we need to modify the set of expired repos from
multiple places, and is also more consistent with how the other
persistor classes work.