Add dnf-langpacks provide
The provide is important when dnf-2.0 is installed and env-goup
"Fedora Workstations" is installing. It requires dnf-langpacks that is obsoletes
by dnf-2.0 but not provided. The patch also change format of obsoletes version
where it use format with only version instead of version-release where it could
interfere with fedora tags in release.