Do not always overwrite the name with the repo ID
The commit 49c7438d3e6de36414f66bc29f717b63dd491571 made it so that
the repo ID was always used, even when the repo name does exist,
because the repo ID was always overwriting the passed in full name.

The value is already set in '_build_repo()' in read.py anyway,
and in the event there is nothing set there, we will fall back
to 'unknown'.

Closes: #674
Approved by: mluscon