Add autocheck_running_kernel config option
It is part of dnf-yum compatibility. This configuration option checks whether there is installed newer kernel with security updates that is not running. Point of this option is to give a hint to sysadmin that maybe they forgot to reboot to the latest kernel. This option changes output of commands 'check-update' and 'updateinfo'.