Dnf timers should require systemd network target(RhBug:1247083)
Dnf timers will run even if there is no running Network manager. 'network-online.target' is a target that actively waits until the nework is "up", where the definition of "up" is defined by the network management software. This means that all timers can still run if there is no internet connection, but there needs to be running network manager.