Add the --forcearch CLI flag
This flag is synthetic (it doesn't set a configuration value
directly). Instead, when this flag is activated, it sets the 'arch'
and 'ignorearch' configuration options. This argument takes a
string value that is one of the value architectures. We enforce that a
valid architecture is specified by using the 'choices' parameter.

RhBug:1427604