Enhance --duplicates for remove command (RhBug:1465292)
From duplicates it tries to reinstall the newest package (it also installs
missing dependencies of that package), removes other, and also us able to remove
protected packages if the newest is available.

Also the transaction doesn't have a power to remove dependent packages (no
allowerasing for the transaction by default).

https://bugzilla.redhat.com/show_bug.cgi?id=1465292