Change code for group conditional packages (RhBug:1427144)
It will use rich deps to solve if conditional package has to be installed.
It will solve the problem if you install pkg in same transaction with install of
group that has conditional package that requires that package. Unfortunately
it works if self.conf.multilib_policy == "best". If it is set to "all" it
provides same problematic behavior.
Additional it is possible to expect different behavior because the package name
was replaced by provide (rich dep) that can lead to install packages that
provide the comps package name but has different package name.

https://bugzilla.redhat.com/show_bug.cgi?id=1427144