cli,demand: Add `root_writable` demand, teach `install` to use it
This is a proposal, interested in feedback.  Currently `dnf` gets
pulled into Fedora Atomic Workstation due to dependencies, but
`rpm-ostree` is the hybrid image/package system that does host updates.

That said there's a whole big issue here we have:
https://github.com/projectatomic/rpm-ostree/issues/405

Basically we're talking about trying to redirect individual commands,
for example having `yum install foo` redirect you to `yum-img layer`
or something (where the latter would be `rpm-ostree install`).

There's a lot of things we could do if we tried to build this bridge;
for example rpm-ostree doesn't implement `search`, and I would be totally
fine delegating that to `dnf`.  (A lot of this is ideally shared in
libdnf but that's its own big topic)

Anyways this is a WIP proposal patch for feedback.

(BTW one interesting side note; if `ostree admin unlock` is used,
 then `dnf` returns to fully writable, just in the overlayfs on `/usr`
 mostly)