Restore 'strict' choice for group installs (#1461539)
This makes the `strict` parameter to `group_install` mean
something again.

When it is set `True`, DNF will behave as yum previously did,
which has been a request for several years now. See:

https://bugzilla.redhat.com/show_bug.cgi?id=1292892
https://bugzilla.redhat.com/show_bug.cgi?id=1337731
https://bugzilla.redhat.com/show_bug.cgi?id=1427365
https://bugzilla.redhat.com/show_bug.cgi?id=1461539

And commits:

91f9ce98dbb630800017f44180d636af395435cd
1e1901822748b754d038dd396655c997281a7201

We have been around the houses on this multiple times. @mikem23
actually got things right way way back in #1292892:

"yum-deprecated behaves more sanely here.

In case 1 [package exists but has dependency issues], it reports
the missing dependency and exits with an error

...

In case 2 [package cannot be found at all], it exits cleanly, but
prints a warning."

Note that this applies to yum's *default* behaviour: it had an
option, --skip-broken, which caused it to skip packages with
dependency errors.

That's exactly what we've wanted all along, and it's also what
1461539 requests. Unfortunately, the initial commit intended to
fix #1292892 - that's 91f9ce9 - did not quite behave as Mike
requested. It treated any 'mandatory' package having dependency
errors *or* not existing at all as fatal errors, while not
treating non-existence *or* errors on the part of a 'default'
or 'optional' package as a fatal error. This introduced *two*
differences from yum's behaviour (making the comps 'type' matter,
which it never did to yum, and failing on non-existence), which
I think was the source of quite a lot of confusion.

We then wound up filing bugs on this - like #1337731 - and the
response to that was 1e19018, which causes dnf to treat *neither*
case as a fatal error for *any* package, which is where we stand
at present.

When resolving a transaction, currently, if it cannot find a
package corresponding to an entry in a group, dnf will log a
warning and carry on. This is good and fine and just what we want
it to do. However, when a package is available but cannot be
installed, it does the wrong thing. The various forks of this
`trans_install` internal function all wind up passing
`optional=True` to the Goal, which tells it that it's fine to
just leave the package out if its dependencies can't be resolved.
Additionally, if this happens, `resolve()` itself does not log
anything.

If you try a group install through the CLI, it will tell you
about packages that have been left out due to dependency errors,
via the `_skipped_packages` function in `dnf/output.py` which
actually sucks them out of the Goal instance itself. But that's
no use to anything besides the CLI. Other things which use
`resolve()`, like anaconda, have no means of accessing this
information, so when this happens to them, the omission of the
package is not logged in any way at all, and they have no way to
find out about it.

If `strict` is set to `False`, the current behaviour described
above will be used: non-installable packages will be skipped. We
implement this by restoring the old separation between `install`
and `install_opt` which was removed in 1e19018. When group_install
is called with `strict=True` we add the packages to the `install`
set; when it's called with `strict=False` we add the packages to
the `install_opt` set. Then `_finalize_comps_trans` requests
strict behaviour for `install` and non-strict for `install_opt`.

It continues to be the case that the comps 'type' - mandatory,
default, or optional - does not matter; the same behaviour will be
used for all three. Again, this is how yum behaved (in the
absence of --skip-broken)

The comps 'types' were *never* intended to dictate yum/dnf
behaviour in terms of whether it should fail on error or not, as I
understand it; they were in fact related to installer UI behaviour.
In the old installer UI, after you selected a group, you would see
all the packages in the group listed. 'mandatory' packages would
be pre-checked for installation and the check would be greyed out
- you couldn't unselect them. 'default' packages would be
pre-checked for installation, but you *could* de-select them if you
wanted. 'optional' packages would not be pre-checked for
installation, but you could *select* them if you wanted. This was
the intent of the comps 'type', as I understand it.

Importantly, a package not existing at all will continue to *not*
be a fatal error whatever `strict` is set to. as this is handled
*before* we reach the `trans_install` function.

Signed-off-by: Adam Williamson <awilliam@redhat.com>