Introduce new dependencies to the spec file
THe Python API for Unbound is required for the DNSSEC extension to work,
but it is optional. The second dependency is python2-enum34 which
backports support for enumerations into Python2. An enumeration is used
for expressing validity in a typed way as opposed to plain integer
constants. This type is then used in type hints, which are written in
Python2 compatible way (comments under function definitions).