[conf] Use VectorString directly rather than converting it to a list.
Converting VectorString to a list resulted in a copy that did not propagate
changes back to the underlying C++ object. A proxy object almost worked,
but it was more a hack than a proper solution. DNF now relies on libdnf
providing additional methods in the SWIG layer that were previously
missing and caused that VectorString did not behave as a regular list.