Support long and short localized messages (RhBug:1585509)
Depending on the context and on the language some messages may be too
long to display on the screen. The translators should be able to provide
the short and long versions of their messages. The original case was
that the word "Size" is translated to Swedish as "Storlek" which is
too long to display in a 5 characters long field. There are more messages
which may have the same problem and their size limit is often calculated
dynamically so it is even more unpredictable.

https://bugzilla.redhat.com/show_bug.cgi?id=1585509