Add basic integration with %_pkgverify_level
RPM 4.14.2 introduced a new low-level security policy for package
verification configured with the %_pkgverify_level macro:

http://rpm.org/wiki/Releases/4.14.2

In DNF, signature verification is done via RPM but in a separate step
that precedes the transaction itself (BaseCli.gpgsigcheck()).  We can
make use of that to catch signature errors and/or import public keys the
same way as if gpgcheck was enabled from the start.  To that end, this
commit forces the gpgcheck and localpkg_gpgcheck options to True if the
policy would result in signature verification anyway.

Resolves RHEL-8 bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1614351