Turn off abbreviations for the command line argument parser
The abbreviations create ambiguity in possible argument options that can
be used and it's an undocumented feature. Disable the abbreviations.
This can potentially break someone's scripts if an undocumented
abbreviation of a command line option was used. In that case, the user
should fix their script according to the documentation.

https://docs.python.org/3.5/library/argparse.html#prefix-matching

The switch is only supported and used if python version is >=3.5, for
lower versions of python the abbreviations can't be turned off and will
still work.