[spec] Add internal yum_compat_level macro
In essence, the yum compat subpackage is built in different flavors,
depending on the distro.  So, instead of checking the distro in multiple
places, let's detect it once, choose the right compat level and then
only use that to filter the files to include/exclude in the subpackage.

The main benefit of this commit is to allow for building the same
subpackage name in different compat levels (see the next commit).