Fix multilib obsoletes (RhBug:1672947)
When a multilib package was installed and it's upgrade was split
into several sub-packages and they all obsoleted the originally
installed packages, DNF did not compute history db data correctly.
The 32bit version of the package was considered as obsoleting
package rather than an upgrade. For that reason, it wasn't stored
in the database and the transaction failed on following error:
"TransactionItem not found for key".