[doc] Change description of skip_if_unavailable option
skip_if_unavailable option is now not only a repo option but also option
of the main config. The commit provides a enhanced description.

Additionally there is a change in default. Formerly it was
skip_if_unavailable=true, but it was changed to False to provide a
better security to users. The commit also change the description
according to new default.