Set log file encoding, similar to setup_stdout()
I have been reviewing this for the dnf plugin etckeeper-dnf, which wants to
log etckeeper output.  (And stop directly interfering with stdout of the
host process, because it breaks Ansible).

I noticed a problem.  Dnf sets up the log file without specifying an
encoding.  In PY3, this means the log file uses the locale encoding.

I believe etckeeper-dnf would not cause a problem here.  However if a
character which cannot be encoded in our locale gets into a log message
in some other way - e.g. from a filename - PY3 will generate
UnicodeEncodeError, not include that log message, and continue. (libdnf
sets logging.raiseExceptions to False).  Let's be more robust, and just
replace the problem characters.

I believe this change is also safe for PY2.  The reason is that dnf.logging
already sets up the logger using a unicode formatting string.  As a result,
log messages must already be either unicode, or ASCII bytestrings.

(Otherwise - by default, when you do not specify an encoding in PY2, the
logging handler accepts both arbitrary bytestrings, and unicode which it
encodes as utf-8.)