[cli] call setup_stdout() ASAP
Move the call to dnf.i18n.setup_stdout(), so it comes before anything
that prints a translated message.  Like dnf.conf.Conf().

This was a problem for py2 only.  setup_stdout() exists because py2 does
not automatically use the locale encoding for stdout, if stdout is not a
tty.

I notice that "import dnf.i18n" automatically calls setup_locale(),
because it loads translations at import time.  If any other module was
similarly crass, and it triggered a translated error message when imported,
it would be inviting a UnicodeEncodeError.  Happily, it looks like there is
no such code path in dnf.  I could have wedged a call to setup_stdout()
within the import lists, to be on the safe side, but that looks unclean to
me.