[cli] call setup_stdout() ASAP
Move the call to dnf.i18n.setup_stdout(), so it comes before anything
that prints a translated message.  Like dnf.conf.Conf().

I notice that "import dnf.i18n" automatically calls setup_locale(),
because it loads translations at import time.  If any other module was
similarly crass, and there was a translated error message, it would be
inviting Mr UnicodeEncodeError.  But it looks like we're safe here.  So
I preferred not to wedge a call to setup_stdout() in the middle of the
import list.

This problem applies to py2 only.  py2 does not use the locale encoding
for stdout, if it is not a tty.