dnf.plugin.log is not used, remove references to it
dnf.plugin.log is not mentioned anywhere.  This includes the logfile names
in const.py.in.  It is not mentioned in dnfpluginscore either.  I have
used etckeeper-dnf which uses dnfpluginscore.logger on my Fedora 29 system.
I have no such file /var/log/dnf.plugin.log.

Remove it from the RPM spec file and the logrotate configuration file.
It is not mentioned anywhere else in the Git history.