Rename man page from dnf.automatic to dnf-automatic
The application name is 'dnf-automatic'. The man page should match that
command name and be available as 'man dnf-automatic'.

Exposing it to users as 'man dnf.automatic' is weird, surprising and
makes it harder to discover.

Even the RPM package name is dnf-automatic so the man page of
dnf.automatic sticks out as unexpected.

Closes: #1387
Approved by: Conan-Kudo