Empty lock file (RhBug:1581824)
If a machine loses power unexpectedly (e.g. in HA configurations)
and the lock is taken by a running dnf process but before the data
(i.e. the pid) can be flushed to the file, then the lock file can be
recreated from the filesystem journal when the machine is powered on again,
but the file is empty.

https://bugzilla.redhat.com/show_bug.cgi?id=1581824

Signed-off-by: Jan Beran <jberan@redhat.com>