Reuse empty lock file (RhBug:1581824)
Fixes a situation where an empty dnf lock file is present
on the file system at dnf startup. The empty lock file is
automatically reused instead of stopping dnf.

Example of creation empty lock file:
If a machine loses power unexpectedly and the lock is taken by a running
dnf process but before the data (i.e. the pid) can be flushed
to the file, then the lock file can be recreated from the filesystem
journal when the machine is powered on again, but the file is empty.

https://bugzilla.redhat.com/show_bug.cgi?id=1581824