[spec] Fix-up yum conflict
In commit 519593cb, I dropped the conflict on "yum" for fedora >= 31 (so
that I could rename the "dnf-yum" subpackage to "yum" in follow-up
commits).

However, I accidentally removed the outer %if macro as well.  As a
result, we now conflict with the stock yum-3 on RHEL-7.  Not nice.

This commit puts back the outer %if to fix that.