Print rpm error messages during transaction (RhBug:1677199)
This patch lets the rpm error messages to be passed on to the user.
Original code basically considered all rpm errors as "Failed to obtain
the transaction lock", which is obviously not always the case. Now dnf
prints out original rpm messages so that user could find out what the
real problem is.

https://bugzilla.redhat.com/show_bug.cgi?id=1677199