[doc] Describe a behavior when plugin is removed (RhBug:1700741)
When plugin is removed or obsoleted all hooks of removed plugins after
transaction are skipped. The behavior was added because some
of dependencies of the plugin could be also removed and unavailable in
memory. We apply the same behavior also for obsoleted plugins, because
we suggest that obsoleting plugin could have a different code base.

The behavior is not applicable for downgrades or upgrades.

https://bugzilla.redhat.com/show_bug.cgi?id=1700741