Remove reset_nice feature
I do not understand why we would want such a feature. In default setting
the nice value is 0 and this code does nothing. When user runs dnf with
"nice" to adjust niceness, he probably has a reason to do so and dnf
should not tamper with value set intentionally by user.