[doc] man page: fix description of Specifying Packages
The description of NEVRA matching in Specifying Packages was plain wrong
with regards to the order of NEVRA forms and lacking in other details.

The new description should be more precise and also cover matching
globs, along with defining how globs are handled in case a glob is used
over multiple NEVRA forms.

I've removed the paragraph about selecting the highest version of a
package and the suitable architecture etc., as that is highly specific
to the install/upgrade/... commands and multiple other commands behave
differently in this regard.

Closes: #1454
Approved by: j-mracek