[doc] man page: remove <package-installed-spec>
<package-installed-spec> was used to denote a command operates on
installed packages only. However, the commands aren't consistent in that
some actually support the <package-spec> variant and others only support
<package-name-spec> variant.

The fact that the commands operate on installed packages is clear from
the context, so I've removed the <package-installed-spec> completely and
used the correct variant of the (non-installed) spec identifier in the
commands in question.