Remove redundant lines in dnf.rpm.log (RhBug:1739457)
RPM element was reported before action started in order to know which
element is in the progress. See RhBug:1305232. Also it was logged
when transaction element action finished and it results in double entry.

Because we cannot modify API callback.filelog() (using additional key
argument) and want to keep information about what element is now process
(in case of frozen system), the removal of call callback.filelog() after
action finished is the only choice.

https://bugzilla.redhat.com/show_bug.cgi?id=1739457