[repoquery] Do --latest-limit filter sooner (RhBug:1561795)
Move --latest-limit filter execution closer to --avaliable, --installed,
--recent etc.
This gives dnf repoquery ability to operate only on the newest available
versions of packages when for example searching packages that require
some capability.

Before the patch "dnf repoquery --latest-limit=1 --whatrequires
CAPABILITY" command first limited the packageset to those providing the
capability and then print only the newest versions of found packages.
Now first the packageset is limited to the latest versions and than the
capability is looked for.

https://bugzilla.redhat.com/show_bug.cgi?id=1561795