logging: always log the repo loading error, but not to stderr (RhBug:1741442)
When loading a repository fails, an exception is thrown, which is logged
(as well as printed to stderr) in dnf/cli/main.py when run from the
command line. When run from API, the exception is thrown but not logged.

We should log these important errors even if we raise an exception on
the API. The problem is the "dnf" logger automatically prints everything
loggged as WARNING or higher to stderr, which means on command line the
error would be printed to stderr twice.

This commit adds another logger, "dnf_no_stderr", which only logs to the
file and not to stderr. This logger is then used to log the repo loading
error.

https://bugzilla.redhat.com/show_bug.cgi?id=1741442